Refactor: 리뷰 반영
- buttons/post 폴더 삭제
- Post component 내에 계층적 구조 -> 병렬적으로
- component 명과 파일 명 통일
- ProfileUserNameButton -> UserNameButton으로 이름 수정
