+++ frontend/src/components/PostContent/index.tsx
