+++ frontend/src/components/PostContent/style.ts
