+++ frontend/src/components/PostImage/style.ts
