+++ frontend/src/components/PostLike/style.ts
