+++ frontend/src/components/PostProfile/index.tsx
@@ -1,6 +1,6 @@
 import { Row } from 'src/components/Grid';
 import ProfileImageButton from 'src/components/buttons/ProfileImageButton';
-import ProfileUsernameButton from 'src/components/buttons/ProfileUsernameButton';
+import UsernameButton from 'src/components/buttons/UserNameButton';
 
 import { Wrapper } from './style';
 
@@ -8,8 +8,8 @@ function Profile() {
   return (
     <Wrapper>
       <Row justifyContent='flex-start'>
-        <ProfileImageButton href='users/123' imgSrc='/images/logo.svg' />
-        <ProfileUsernameButton href='user/123' userName='tiger' />
+        <ProfileImageButton href='users/123' imageSrc='/images/logo.svg' />
+        <UsernameButton href='user/123' userName='tiger' />
       </Row>
     </Wrapper>
   );
