+++ frontend/src/components/PostProfile/style.ts
