+++ frontend/src/components/Timeline/index.tsx
@@ -6,7 +6,7 @@ import { Wrapper } from './style';
 function Timeline() {
   return (
     <Wrapper>
-      <Card width={600} height={500}>
+      <Card width={600} height={600}>
         <Post />
       </Card>
     </Wrapper>
