+++ frontend/src/components/buttons/LikeButton/index.tsx
