+++ frontend/src/components/buttons/LikeButton/style.ts
