+++ frontend/src/components/buttons/ProfileImageButton/index.tsx
@@ -6,15 +6,15 @@ import { Wrapper } from './style';
 
 interface Props {
   href: string;
-  imgSrc: string;
+  imageSrc: string;
 }
 
-function ProfileImageButton({ href, imgSrc }: Props) {
+function ProfileImageButton({ href, imageSrc }: Props) {
   return (
     <Wrapper>
       <Link href={href}>
         <a href={href}>
-          <Image src={imgSrc} width='36' height='36' />
+          <Image src={imageSrc} width='36' height='36' />
         </a>
       </Link>
     </Wrapper>
@@ -23,7 +23,7 @@ function ProfileImageButton({ href, imgSrc }: Props) {
 
 ProfileImageButton.propsType = {
   href: PropTypes.string.isRequired,
-  imgSrc: PropTypes.string.isRequired,
+  imageSrc: PropTypes.string.isRequired,
 };
 
 export default ProfileImageButton;
