+++ frontend/src/components/buttons/slides/Common/index.tsx
