+++ frontend/src/components/buttons/slides/Common/style.ts
