+++ frontend/src/components/buttons/slides/LeftSlideButton/style.ts
