+++ frontend/src/components/buttons/slides/RightSlideButton/style.ts
